iommu/vt-d: Fix bugs when doing dmatest in VM with scalable-mode vIOMMU
First bug is that the dmar_domain->subdevices is not initialized.

Second bug is that the dmar_domain->lock set in intel_nested_iotlb_sync_user.
In fact, the next level functions will set the same lock. So, remove
the lock in nested.c

The last bug is checking the nested_users to add/del info.

Signed-off-by: Yi Sun <yi.y.sun@linux.intel.com>