platform/x86/intel/ifs: ArrayBIST trigger from 1 sibling
During poweron testing core team informed us that unlike SAF, Array
test needs to be triggered from  only a single sibling thread of the
initating core.

We were also seeing lot of retries if the sibling of the core initiating
the test is in m-wait state. As a mitigation we spin the sibling until
the first thread completes the test.

Additional details in below HSDs.

[1] https://hsdes.intel.com/resource/14018157349
[2] https://hsdes.intel.com/resource/16018961627
[3] https://hsdes.intel.com/resource/13010567499

Signed-off-by: Jithu Joseph <jithu.joseph@intel.com>