PCI: Use pci_dev_wait() in pci_bridge_wait_for_secondary_bus()
Currently we wait for an additional 100 ms if the device does not
respond after reset. However, that seems not to be enough for certain
Intel Titan Ridge xHCI USB controllers. If a device is plugged in during
s2idle the controller may take longer than that to be able to respond to
PCI config requests. So instead of waiting for additional 100 ms use
pci_dev_wait() to figure out when the device is ready.

Link: https://bugzilla.kernel.org/show_bug.cgi?id=216728
Reported-by: Chris Chiu <chris.chiu@canonical.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>