KVM: TDX: make prefault EPT as the default MapGPA behavior for TDXIO
The prefault parameter for MapGPA is not defined in
Guest-Host-Communication Interface (GHCI) spec.

The prefault is necessary if EPT is shared with IOMMU, like in TDXIO.
Host should enforce this implementation in such case.

For now, just force enable prefault for TDXIO. Later may expand it for
generic use cases.

Signed-off-by: Xu Yilun <yilun.xu@intel.com>