Revert "msi: export msi_domain_alloc_irqs() and msi_domain_free_irqs() symbols"
This reverts commit daef3777b96726aa0117d5c5abe3a157ffd8ff97.

This commit formed part of the original device MSI solution
that is made obsolete by the IMS support in v6.2.

The commit being reverted was not accepted upstream. This
change is specific to the BKC.

In order for revert to succeed this also reverts a compile fix
made to the code being reverted:

Remove fixup introduced in f2bc6f0ae8cf ("vfio: idxd: Fix a
few compilation issues").

Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>