Revert "msi: Fix an issue of removing msi_irqs group"
This reverts commit a3ce582397b66f6634b6feb1a07759f75a91a7a2.

This fix is to obsolete code that will be reverted in subsequent
change.

In order for revert to proceed also undo compile fixes done in
f2bc6f0ae8cf ("vfio: idxd: Fix a few compilation issues").
The compile fix revert is squashed with the revert to keep
kernel bisectable.

Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>