Revert "x86/msi: Rename and rework pci_msi_prepare() to cover non-PCI MSI"
This reverts commit 521e65cc63d3c694a25b99b68dc39f98f33b7d3e.

This commit formed part of the original device MSI solution
that is made obsolete by the IMS support in v6.2.

The commit being reverted was not accepted upstream. This
change is specific to the BKC.

In order for revert to succeed, also revert a compile fix made to
the reverted code: f2bc6f0ae8cf ("vfio: idxd: Fix a few
compilation issues").

The two changes are reverted together to keep the
kernel bisectable.

Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>