vfio: Port of common lib code to new API
This is a fixup of commit dad35ec4b8fe ("vfio: common lib code for
setting up Interrupt Message Store") to port it to the new v6.2 APIs.

This is PoC code created specifically in support of the BKC kernel.
The code should be reworked to contain where interrupts
are allocated and also to route the pci dev in an integrated way.
Interrupt allocation should be changed to be similar to the new
dynamic MSI-X allocation.

NOT FOR UPSTREAM

Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>