vfio: Create and move to new utilities for IMS info
vfio needs a way to obtain the information associated with
the IMS interrupt backing a particular MSI-X interrupt. Given
an MSI-X interrupt it needs to be able to query the IMS
interrupt backing it as well as that interrupt's Linux
IRQ number.

The original utilities providing this information are
no longer usable (and about to be removed) since it
relied on an implementation of IMS support that has been
replaced with the interrupt subsystem changes found in
v6.2.

Create new utilities and move existing usages to it.

This is a PoC patch created specifically for the BKC kernel.

NOT FOR UPSTREAM.

Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>