vfio: Do not program pasid after interrupt allocation
Previous IMS implementation allowed PASID of interrupt to
be changed after interrupt allocation. In IMS support found
starting in v6.2 the PASID is associated with interrupt
when interrupt is allocated and cannot be changed.

Usage of old APIs are disabled. The usages need to be
moved to new core IMS support.

This is a PoC patch created specifically for BKC.

NOT FOR UPSTREAM

Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>