x86/mce: Check if address in machine check bank is virtual
Intel machine check architecture allows the CPU to supply various types
of addresses in the IA32_MCi_ADDR register. The type is specified in the
IA32_MCi_MISC register.

Historic CPU implementations always supplied a physical address for
recoverable errors and Linux did not include a check for the type
of the address.

Add code to check if the address type is virtual. Pass the result
to the function called by task_add() via a new flag bit in the
task structure.

A subsequent change will handle the virtual address case.

Signed-off-by: Tony Luck <tony.luck@intel.com>