Fix bug for TDX init in guest: correct cpus_mask
Previous code use a pointer in order to save the mask for later
use. But after the cpus_mask changed, the content pointed by that
pointer is also changed. So this leads to incorrect cpus_mask be
used for recover after other settings done.

As the result, incorrect CPU affinity is set for process
"init" and all descendent processes inherit this wrong affinity
as default (i.e. they will only run on boot core by default).

The fix is simple that just save the value of the cpus_mask so it
can be used correctly for later restore.

Signed-off-by: zhongjie <zhongjie.shi@intel.com>