KVM: MSRLIST CPUID Enabling
This patch aims to expose MSRLIST NIs to guest so that guest could
utilize MSRLIST as an optimization for MSR save/restore. The RDMSRLIST
and WRMSRLIST are instructions to read/write a software-provided list of
up to 64 MSRs to/from memory in a performance manner.

The virtualization behavior mainly includes VM exit. MSRLIST allows KVM to
emulate multiple MSRs in one VM-exit and then it can reduce the number of
VM-exit if 2 or more MSRs to access are intercepted, The difference between
RDMSR/WRMSR and RDMSRLIST/WRMSRLIST is that the former results are put in GPRS,
but the latter results are are put in guest memory.

The MSRLIST instruction is currently documented in the latest "instruction
set extensions" manual (ISE)[1].

[1] https://cdrdv2.intel.com/v1/dl/getContent/671368

Signed-off-by: Dan Wu <dan1.wu@intel.com>