REVERTME: x86: pci-tdisp: disable heartbeat & keyupdate
spdm_owner function should be redesigned, if heartbeat and keyupdate
happen concurrent, the first request of them will change owner to kernel
which the second request is handled by agent, need to guarantee spdm
owner would not be changed until all request are done in that period.
Currently, just skip heartbeat and keyupdate for integration progress.

Signed-off-by: Ming Li <ming4.li@intel.com>