selftests: restrictedmem: Check hugepage-ness of shmem file backing restrictedmem fd
For memfd_restricted() calls without a userspace mount, the backing
file should be the shmem mount in the kernel, and the size of backing
pages should be as defined by system-wide shmem configuration.

If a userspace mount is provided, the size of backing pages should be
as defined in the mount.

Also includes negative tests for invalid inputs, including fds
representing read-only superblocks/mounts.

Signed-off-by: Ackerley Tng <ackerleytng@google.com>
Link: https://lore.kernel.org/r/0061b62966d34952fb9f51235d31100df0baf450.1680306489.git.ackerleytng@google.com
Signed-off-by: Binbin Wu <binbin.wu@linux.intel.com>