x86/microcode: When preparing for NMI update initialize common atomics before bailout
prepare_for_update() was refactored for accomodate the NMI handling and
selectively not doing when UNIFORM update is enumerated for Intel CPUs.
Possibly its also needed for AMD, since the claim it AMD isn't subject to
the same issues like Intel CPUs.

Initialize the cpus_in and cpus_out atomics since they are required even
when NMI isn't in play.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>