KVM:x86: Store host S_CET msr data to vmcs field before vm-entry
Store host S_CET msr to vmcs field manually before vm-entry. By doing so,
if exit_load_host bit is set then the stored host S_CET value can be
correctly load to the MSR after vm-exit. If not store the host value to
vmcs field and exit_load_bit is set, the HW actually restore the default
value in vmcs field to the MSR, which is 0 in this case.

Signed-off-by: Yang Weijiang <weijiang.yang@intel.com>