x86/microcode/intel: Save early microcode if it matches BIOS FIT load
This commit was introduced to fix the APs failing to find the new microcode patch,
but it turns out the root cause for the issue seems to be a hardware quirk where
even if the `IA32_MCU_ENUMERATION` doesn't explicitly show support for uniform update
with package scope but the platforms may support a wider scope.

```
From the spec:
Some processors may support Uniform Microcode Update even if it is not enumerated,
or may support Uniform Microcode Update at a wider scope than what is enumerated.
This is known as model-specific Uniform Microcode Update and may occur when certain
features are enabled. For example, when Intel® Software Guard Extensions (Intel® SGX)
is enabled, the CPU will enable platform scoped Uniform Microcode Update even if it
is not enumerated or is enumerated with a narrower scope. 
```

On the test platform we do see SGX enabled and since all the APs didn't find a newer
microcode, returned patch NULL as expected. Tried disabling SGX on the test platform,
and as expected all the APs could properly find the new microcode patch and we able to
apply during early load.

Secondly, for future rollback cases, we already save the patch even if the current
CPU revision and the new microcode revision are the same. So the current change
will not be needed even in this case.

This is what I see with SGX disabled on APs

```

[    1.764330] microcode: has_newer_microcode: mc_hdr->rev = 0xd000389, new_rev = 0xd0001e0
[    1.764330] microcode: scan_microcode: patch is NULL, and microcode rev is newer than current
[    1.764330] microcode: Setting patch at 0x(____ptrval____) val = 0xff218c727ea84ab0

```

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>