ice: fix ice module unload
Disabling interrupts before the PFR reset in the device removal routine
(ice_remove()), was the cause of multiple errors when unloading ice module.
In this patch, the interrupts disablement was moved from the device
deinitialization subprocedure to appropriate places in particular routines.
In the case of ice_remove() it was placed after the PFR reset.

Additionally, this patch fixes the interface reload using devlink, by
clearing the interrupt scheme after the PFR reset in the loading routine.
Before that, there was a failure on interrupt scheme initialization.

Fixes: 95e2130979fa ("ice: split probe into smaller functions")
Signed-off-by: Jakub Buchocki <jakubx.buchocki@intel.com>