ice: add q_vector->vf_reg_idx for configuring VF interrupts
The ice_vc_cfg_irq_map_msg function processes the
VIRTCHNL_OP_CONFIG_IRQ_MAP VF message. This message is a request form the
VF to program the mapping between interrupt vectors and queues.

The driver must map the relative index coming from the VF into an index for
use in programming the relevant hardware registers.

For Single Root IOV, this is straight forward, because the associated
registers also use a relative index. Thus, the current code simply passes
the vector_id from the VF directly into ice_cfg_interrupt() and in turn
into ice_cfg_txq_interrupt() and ice_cfg_rxq_interrupt().

This will not work for Scalable IOV, as the associated registers expect an
index which is relative to the entire vector list in the PF.

To handle this, introduce a new field to ice_q_vector: vf_reg_idx. This
field represents the index to use when programming VF related registers.
The q_vector structure has a 2 byte hole we can fill for this value. To
keep things in a nice order, move the num_ring_rx and num_ring_tx fields
into this hole so that the vf_reg_idx can be placed next to the reg_idx
field.

This value will be identical to q_vector->reg_idx for VSI types other than
ICE_VSI_VF.

For the ICE_VSI_VF type, we instead calculate the appropriate index using
the q_vector->v_idx. However, we must account for the interrupts which are
not associated with queues (and thus not tracked in the q_vector array).
Assign the q_vector->vf_reg_idx inside the appropriately named
ice_calc_vf_reg_idx() function which is already used to calculate the
q_vector->reg_idx. Since we can just assign these to the q_vector, convert
this function to void.

With this change, the ice_cfg_interrupt function is now able to be generic
so that we can use it for both Single Root IOV and the planned Scalable IOV
support.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>