ice: add the ICE_VSI_ADI type for tracking Scalable VFs
A future set of changes will add support for a new type of subfunction
representing a virtual device implemented using the VFIO framework.

The VFIO framework will allow creating a device that is represented to the
VM using a similar register interface as the existing PCIe SR-IOV VFs.
These devices will use the existing iAVF virtual function driver and the
same PF to VF communication mechanism known as virtchnl. We call these
subfunction virtual devices Scalable VFs.

For the VSI that represents this subfunction, we can't easily use an
existing VSI type. The ICE_VSI_VF is tied too heavily to the hardware VF
VSI type. This can't work because these subfunctions do not get represented
in the hardware as full PCIe VFs. The ICE_VSI_SF type represents a netdev
based subfunction. We can't re-use this type because its logic would not
allow adding a VF pointer and hooking into the virtchnl interface.

Instead, introduce a new ICE_VSI_ADI type, based on the hardware VMDQ2 VSI,
(as ICE_VSI_SF and similar VSI types).

This VSI type represents the VSI for the Scalable VF. It is similar to the
ICE_VSI_VF in many ways, but its underlying hardware representation is the
VMDQ2 VSI, and it tracks its hardware resources slightly differently.

Add the VSI type and hook it up to the various VSI initialization and
configuration functions. In most cases it either follows an existing flow,
either the ICE_VSI_VF flow or the ICE_VSI_SF flow.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>