devlink: add VFIO devlink port flavour
The DEVLINK_PORT_FLAVOUR_VFIO port type is intended to represent a portion
of a networking device that is exposed to userspace via a struct
vfio_device. This is similar to a DEVLINK_PORT_FLAVOUR_PCI_VF, which would
represent a PCI SR-IOV VF. The intended use case for this port flavour is
to allow users to dynamically add VFIO-based virtual functions.

I considered reusing DEVLINK_PORT_FLAVOUR_PCI_VF, but decided against it.
This port does not get represented as a normal PCI virtual function,
instead being represented to userspace via a struct vfio_device interface.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>