ice: connect devlink port hw_addr to VF MAC address
Currently devlink subfunctions have a hardware address that is disconnected
from the VF MAC address. This is likely to confuse users who see a MAC
address and expect it to be used as the VF address.

Fix this by refactoring so that the devlink hardware address is shared to
the VF as its MAC address. This aligns with SR-IOV and the VF MAC address
that can be reported via netlink.

Convert vf->hw_lan_addr into a pointer, and modify the SR-IOV
implementation to store it as part of its private structure in
ice_sriov_vf. Have the hw_lan_addr point to this storage. For Scalable VFs,
point to the dyn_port->hw_addr field.

When the devlink address is updated, call the ice_vf_notify_mac_addr()
function to ensure that the VF gets reset if necessary.

With this change, VFs should now use the address provided by devlink
instead of using their own randomized address. VFs are allowed to update
and assign their own address if userspace has not assigned the address
manually, just as in SR-IOV with the netlink VF mac address.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>