ice: implement Scalable IOV VF operations table
The Scalable IOV VF implementation differs from Single Root IOV in a few
key ways. To handle this, the ice_vf_ops table needs to be implemented for
Scalable IOV.

Add the required registers for implementing the Scalable IOV operations,
and then implement the table of operations. Add fields as necessary to the
private ice_scalable_dev structure.

This does not yet register the VF and store it into the PF's hash table.
Thus, these operations are currently unused. A following change will add
code to finish registration of the VF.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>