ice: add devlink parameter to configure Scalable IOV mode
The ice device can support Scalable IOV. However, this support requires two
device features to be enabled, 1) extended PF BAR memory space to allow
space for the Scalable registers, 2) PASID capability support.

These features are not on by default, but can be enabled by writing to the
flash. Add support for a new "scalable_mode" devlink parameter which
reports the current status (whether the device is in scalable mode), and
can be toggled by userspace to reconfigure the device into scalable mode.

This is a "permanent" parameter as it is written to the flash. Because it
changes the PCI config space, a reboot is required for the settings to
actually take effect.

Older versions of the flash contents do not support configuring these
parameters. On those systems, the request to read the NVM feature will fail
with Admin queue error ICE_AQ_RC_ENOENT. In this case, we translate the
error to -EOPNOTSUPP which effectively disables the devlink parameter
entirely. This is done rather than reporting the value as "disabled" since
it makes clear that it is not possible to enable the support at all.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>